* last bits of delta..

+++ branches/pisi-delta/pisi/atomicoperations.py
@@ -296,7 +296,22 @@ class Install(AtomicOperation):
                 oldconfig = path + '.old'
                 if os.path.exists(newconfig):
                     os.unlink(newconfig)
-                os.rename(path, newconfig)
+
+                # In the case of delta packages: the old package and the new package
+                # may contain same config typed files with same hashes, so the delta
+                # package will not have that config file. In order to protect user
+                # changed config files, they are renamed with ".old" prefix in case
+                # of the hashes of these files on the filesystem and the new config 
+                # file that is coming from the new package. But in delta package case
+                # with the given scenario there wont be any, so we can pass this one.
+                # If the config files were not be the same between these packages the
+                # delta package would have it and extract it and the path would point
+                # to that new config file. If they are same and the user had changed 
+                # that file and using the changed config file, there is no problem 
+                # here.
+                if os.path.exists(path):
+                    os.rename(path, newconfig)
+
                 os.rename(oldconfig, path)
 
         # Delta package does not contain the files that have the same hash as in
